Add registration commands, rename global commands
