+++ tests/cypress/fixtures/registration.json
@@ -1,5 +1,4 @@
 {
-  "account": "fo",
   "given_name": "Cypress_Name",
   "family_name": "Family_Name"
 }
