+++ tests/cypress/fixtures/registrationPO.json
@@ -1,4 +1,3 @@
 {
-  "account": "po",
   "company_name": "Cypress_Company_Name"
 }
