+++ tests/cypress/support/e2e.ts
@@ -1,4 +1,5 @@
-import './commands'
+import './commands/global'
+import './commands/registration'
 import './triggers/before'
 import '@frsource/cypress-plugin-visual-regression-diff'
